requirements.txt
altgraph==0.17.5
black==25.12.0
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==2.0.0
click==8.3.1
cssselect2==0.8.0
defusedxml==0.7.1
Flask==3.1.2
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown2==2.5.4
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pdoc==16.0.0
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pytest==9.0.2
pytokens==0.3.0
setuptools==80.9.0
tinycss2==1.5.1
webencodings==0.5.1
Werkzeug==3.1.4
